Working edit remove org user

+++ backend/app/routes/users.py
@@ -6,7 +6,9 @@ from backend.app.utils.user import UserWithRole
 from backend.common.core.enums import UserRole
 from backend.common.db.init import get_session
 from backend.common.models.user import User
+from backend.common.models.organization import NewOrganization
 from backend.common.repositories.users_repo import users_repo
+from backend.common.repositories.organizations_repo import organizations_repo
 
 router = APIRouter(
     prefix="/users",
@@ -49,7 +51,36 @@ async def update_user(
     session: AsyncSession = Depends(get_session),
     user: User = Depends(pub_admin),
 ):
+    if user_updates.id == user.id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Invalid request",
+        )
     return await users_repo.update(session, user, user_updates)
+
+
+@router.post("/{id}/remove", response_model=User)
+async def remove_user_from_org(
+    id: int,
+    session: AsyncSession = Depends(get_session),
+    user: User = Depends(pub_admin),
+):
+    if id == user.id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Invalid request",
+        )
+    existing_user = await users_repo.get(session, id, user)
+    if existing_user is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+    # TODO: we should have a concept of a personal org that gets reassigned
+    created_org = await organizations_repo.create(
+        session, user, NewOrganization(name="Personal")
+    )
+    existing_user.roles = [UserRole.PUBADMIN]
+    return await users_repo.change_org(session, existing_user, created_org.id)
 
 
 @router.delete("/{id}", response_model=None)
@@ -58,4 +89,9 @@ async def delete_user(
     session: AsyncSession = Depends(get_session),
     user: User = Depends(pub_admin),
 ):
+    if id == user.id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Invalid request",
+        )
     return await users_repo.delete(session, user, id)
